+++ assets/globe/render-canvas.js
@@ -12,8 +12,8 @@
 
 import {
   project, splitAtSeam, densify, clipToCap,
-} from './projection.js';
-import { ringsOf, ringsOfRegion } from './worlddata.js';
+} from '../geo/projection.js';
+import { ringsOf, ringsOfRegion } from '../geo/worlddata.js';
 
 const STEP_DEG = 2;